fixed borken inspector when multi-object editing
Somehow I've never in all these years tried to select multiple obejcts ing HashedStrings before. Unity's serializer didn't like that are mangled all of the values due to an improper drawer.

The property drawer is now fixed and HashedStrings will be displayed and serialized correctly when selecting multiple objects using them. Note that the hashed number will still now draw correctly but at least the string data (which is the part that is serialized) will be safe.